pyproject.toml
[project]
name = "league-gotta-bike"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "channels[daphne]>=4.3.1",
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-allauth[mfa]>=65.12.0",
    "django-anymail[resend]>=13.1",
    "django-htmx>=1.26.0",
    "django-phonenumber-field>=8.3.0",
    "django-push-notifications[wp]>=3.2.1",
    "django-tailwind[reload]>=4.2.0",
    "django-tasks>=0.8.1",
    "logfire[asgi,django,psycopg,sqlite3]>=4.12.0",
    "phonenumbers>=9.0.16",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.11.0",
    "resend>=2.17.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.19.0",
    "django-debug-toolbar>=6.0.0",
    "django-reset-migrations>=0.4.0",
    "django-stubs>=5.2.5",
    "pytest>=8.4.2",
    "cookiecutter>=2.6.0", # needed for tailwind app
    "ruff>=0.13.0",
    "ty>=0.0.1a20",
        "honcho>=2.0.0",
]
